# documentation/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Post away App: Customers API",
        "description": "API for managing posts, comments, likes,  and their signup/login operations"
      },
      "host":"localhost:3000",
      "basePath": "/",
      "schemes": ["http"],
      "paths":{
...
        "/api/posts": {
            "get": {
              "summary": "Get all posts of a user",
              "operationId": "userAllPosts",
              "tags": ["Posts"],
              "responses": {
                "200": {
                  "description": "Successfully retrieved user posts",
                  "content": {
                    "application/json": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "number" },
                          "postId": { "type": "number" },
                          "caption": { "type": "string" },
                          "imageUrl":{ "type": "string" },
                          "status":{ "type": "string" }
                        }
                      }
                    }
                  }
                },
                "400": {
                  "description": "User not logged in",
                  "content": {
                    "application/json": {
                      "example": {
                        "status": "Error",
                        "msg": "Please login first!"
                      }
                    }
                  }
                }
              }
            },
            "post": {
              "summary": "Create a new post",
              "operationId": "createPosts",
              "tags": ["Posts"],
              "requestBody": {
                "required": true,
                "content": {
                  "multipart/form-data": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "caption": {
                          "type": "string",
                          "description": "Caption for the post"
                        },
                        "imageUrl": {
                          "type": "string",
                          "format": "binary",
                          "description": "Image for the post"
                        }
                      },
                      "required": ["caption", "imageUrl"]
                    }
                  }
                }
              },
              "responses": {
                "201": {
                  "description": "Successfully created a new post",
                  "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                  "id": { "type": "number" },
                                  "postId": { "type": "number" },
                                  "caption": { "type": "string" },
                                  "imageUrl":{ "type": "string" },
                                  "status":{ "type": "string" }
                                }
                              }
                          }
                    }
                  }
                },
                "400": {
                  "description": "Missing required fields or failed to create post",
                  "content": {
                    "application/json": {
                      "example": {
                        "status": "Error",
                        "msg": "Posting Failed. Please try again!"
                      }
                    }
                  }
                }
              }
            }
        },
        "/api/posts/{id}": {
            "get": {
              "summary": "Get post by ID",
              "operationId": "postById",
              "tags": ["Posts"],
              "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                    "type": "integer"
                  },
                  "description": "ID of the post"
                }
              ],
              "responses": {
                "200": {
                  "description": "Successfully retrieved post by ID",
                  "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                  "id": { "type": "number" },
                                  "postId": { "type": "number" },
                                  "caption": { "type": "string" },
                                  "imageUrl":{ "type": "string" },
                                  "status":{ "type": "string" }
                                }
                              }
                          }
                    }
                  }
                },
                "400": {
                  "description": "Post not found",
                  "content": {
                    "application/json": {
                      "example": {
                        "status": "Error",
                        "msg": "The post with ID 1 does not exist in the system."
                      }
                    }
                  }
                }
              }
            },
            "put": {
              "summary": "Update post by ID",
              "operationId": "userUpdatePost",
              "tags": ["Posts"],
              "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                    "type": "integer"
                  },
                  "description": "ID of the post"
                }
              ],
              "requestBody": {
                "required": true,
                "content": {
                  "multipart/form-data": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "caption": {
                          "type": "string",
                          "description": "Caption of the post"
                        },
                        "imageUrl": {
                          "type": "string",
                          "format": "binary",
                          "description": "Image of the post"
                        }
                      },
                      "required": ["caption", "imageUrl"]
                    }
                  }
                }
              },
              "responses": {
                "200": {
                  "description": "Successfully updated post",
                  "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                  "id": { "type": "number" },
                                  "postId": { "type": "number" },
                                  "caption": { "type": "string" },
                                  "imageUrl":{ "type": "string" },
                                  "status":{ "type": "string" }
                                }
                              }
                          }
                    }
                  }
                },
                "400": {
                  "description": "Failed to update post",
                  "content": {
                    "application/json": {
                      "example": {
                        "status": "Error",
                        "msg": "Post Updation failed!"
                      }
                    }
                  }
                }
              }
            },
            "delete": {
              "summary": "Delete post by ID",
              "operationId": "userDeletePost",
              "tags": ["Posts"],
              "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                    "type": "integer"
                  },
                  "description": "ID of the post"
                }
              ],
              "responses": {
                "200": {
                  "description": "Successfully deleted post",
                  "content": {
                    "application/json": {
                      "example": {
                        "status": "Success",
                        "msg": "Post deleted successfully"
                      }
                    }
                  }
                },
                "400": {
                  "description": "Failed to delete post",
                  "content": {
                    "application/json": {
                      "example": {
                        "status": "Error",
                        "msg": "Something went wrong. Please try again!"
                      }
                    }
                  }
                }
              }
            }
        },
        "/api/posts/caption/{caption}": {
            "get": {
              "summary": "Get posts by caption",
              "operationId": "getPostByCaption",
              "tags": ["Caption"],
              "parameters": [
                {
                  "name": "caption",
                  "in": "path",
                  "required": true,
                  "schema": {
                    "type": "string"
                  },
                  "description": "Caption query string"
                }
              ],
              "responses": {
                "200": {
                  "description": "Successfully retrieved posts by caption",
                  "content": {
                    "application/json": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Post"
                      }
                    }
                  }
                },
                "200": {
                  "description": "No posts found with the caption",
                  "content": {
                    "application/json": {
                      "example": {
                        "status": "Success",
                        "msg": "No post found with given query"
                      }
                    }
                  }
                }
              }
            }
        },
...
    }   
}